docs(typescript): updated documentation about module augmentation
I felt like it made sense to merge "Exact theme type" and "Additional properties in the theme" into a single "Extending the Theme".

This is an initial draft as I haven't tested any of this myself, but based on !1090 and my own experience this is how I expect it will work.